refactor: move stringLocationsFromStream into its own file
Progress toward https://github.com/decaffeinate/decaffeinate/issues/557

index.js is big enough as-is, and I'm hoping to consolidate this code with the
herestring code, so it's nice to split it out.
