+++ src/index.js
@@ -5,6 +5,7 @@ import SourceLocation from './SourceLocation.js';
 import SourceToken from './SourceToken.js';
 import SourceTokenList from './SourceTokenList.js';
 import SourceType from './SourceType.js';
+import stringLocationsFromStream from './utils/stringLocationsFromStream.js';
 import tripleQuotedStringSourceLocations from './utils/tripleQuotedStringSourceLocations.js';
 
 /**
@@ -38,91 +39,6 @@ export default function lex(source: string): SourceTokenList {
     previousLocation = location;
   } while (location.type !== EOF);
   return new SourceTokenList(tokens);
-}
-
-function stringLocationsFromStream(stream: BufferedStream): Array<SourceLocation> {
-  let startOfStringInterpolation = stream.hasNext(DSTRING, INTERPOLATION_START);
-
-  if (!startOfStringInterpolation) {
-    return [];
-  }
-
-  let result = [];
-  let first = stream.shift();
-  let quote = '"';
-
-  result.push(
-    // "abc#{def}ghi"
-    // ^
-    new SourceLocation(
-      STRING_START,
-      first.index
-    ),
-    // "abc#{def}ghi"
-    //  ^
-    new SourceLocation(
-      STRING_CONTENT,
-      first.index + quote.length
-    )
-  );
-
-  let loc;
-  let insideInterpolation = true;
-  while (true) {
-    if (insideInterpolation) {
-      result.push(...stringLocationsFromStream(stream));
-    }
-    loc = stream.shift();
-    if (loc.type === EOF) {
-      throw new Error('Reached end of file before finding end of string interpolation.');
-    } if (loc.type === INTERPOLATION_START) {
-      insideInterpolation = true;
-      result.push(
-        new SourceLocation(
-          loc.type,
-          loc.index
-        )
-      );
-    } else if (loc.type === INTERPOLATION_END) {
-      insideInterpolation = false;
-      result.push(new SourceLocation(
-        loc.type,
-        loc.index
-      ));
-    } else if (!insideInterpolation && loc.type === first.type) {
-      let next = stream.peek();
-      if (next.type === INTERPOLATION_START) {
-        // "abc#{def}ghi#{jkl}mno"
-        //           ^^^
-        result.push(new SourceLocation(
-          STRING_CONTENT,
-          loc.index
-        ));
-      } else {
-        // "abc#{def}ghi#{jkl}mno"
-        //                    ^^^
-        result.push(new SourceLocation(
-          STRING_CONTENT,
-          loc.index
-        ));
-        // "abc#{def}ghi#{jkl}mno"
-        //                       ^
-        result.push(new SourceLocation(
-          STRING_END,
-          next.index - quote.length
-        ));
-        break;
-      }
-    } else {
-      // Anything inside interpolations.
-      result.push(new SourceLocation(
-        loc.type,
-        loc.index
-      ));
-    }
-  }
-
-  return result;
 }
 
 function combinedLocationsForMultiwordOperators(stream: BufferedStream, source: string): Array<SourceLocation> {
